+++ Coimbra/GameObjectUtility.cs
@@ -73,7 +73,9 @@ namespace Coimbra
             }
             else if (CoimbraUtility.IsPlayMode)
             {
+#pragma warning disable COIMBRA0019
                 Object.Destroy(gameObject);
+#pragma warning restore COIMBRA0019
             }
             else
             {