+++ Coimbra/ObjectUtility.cs
@@ -47,7 +47,9 @@ namespace Coimbra
 
             if (CoimbraUtility.IsPlayMode)
             {
+#pragma warning disable COIMBRA0019
                 Object.Destroy(o);
+#pragma warning restore COIMBRA0019
             }
             else
             {