+++ Roslyn~/Coimbra.Roslyn/UnityTypes.cs
@@ -6,6 +6,8 @@
 
         public static readonly TypeString ComponentClass = new("Component", Namespace);
 
+        public static readonly TypeString GameObjectClass = new("GameObject", Namespace);
+
         public static readonly TypeString ObjectClass = new("Object", Namespace);
     }
 }